Add GPU Information menu
