+++ src/renderer/store/message.ts
@@ -1,27 +1,5 @@
+import { Message } from "@/common/message";
 import { reactive, UnwrapNestedRefs } from "vue";
-
-export type ListItem = {
-  text: string;
-  children?: string[];
-};
-
-export type List = {
-  type: "list";
-  items: ListItem[];
-};
-
-export type Link = {
-  type: "link";
-  text: string;
-  url: string;
-};
-
-export type Attachment = List | Link;
-
-export type Message = {
-  text: string;
-  attachments?: Attachment[];
-};
 
 export class MessageStore {
   private _queue: Message[] = [];
